requirements.txt
mod_wsgi
Django==3.0.*
mysqlclient
django-crispy-forms
Pillow